AWS SDK - adding SQS propagation
