+++ instrumentation/apache-camel-2.20/javaagent/apache-camel-2.20-javaagent.gradle
@@ -25,7 +25,7 @@ dependencies {
 
   testInstrumentation project(':instrumentation:apache-httpclient:apache-httpclient-2.0:javaagent')
   testInstrumentation project(':instrumentation:servlet:servlet-3.0:javaagent')
-  testInstrumentation project(':instrumentation:aws-sdk:aws-sdk-1.11:javaagent')
+  testInstrumentation project(':instrumentation:aws-sdk:aws-sdk-1.11-core:javaagent')
 
   testImplementation group: 'org.spockframework', name: 'spock-spring', version: "$versions.spock"
