+++ instrumentation/aws-sdk/aws-sdk-1.11-core/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/awssdk/v1_11/core/AwsHttpClientInstrumentation.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.opentelemetry.javaagent.instrumentation.awssdk.v1_11.core;
+
+import static io.opentelemetry.javaagent.instrumentation.awssdk.v1_11.core.AwsSdkClientTracer.tracer;
+import static io.opentelemetry.javaagent.instrumentation.awssdk.v1_11.core.TracingRequestHandler.CONTEXT_SCOPE_PAIR_CONTEXT_KEY;
+import static java.util.Collections.singletonMap;
+import static net.bytebuddy.matcher.ElementMatchers.isAbstract;
+import static net.bytebuddy.matcher.ElementMatchers.isMethod;
+import static net.bytebuddy.matcher.ElementMatchers.named;
+import static net.bytebuddy.matcher.ElementMatchers.not;
+
+import com.amazonaws.AmazonClientException;
+import com.amazonaws.Request;
+import com.amazonaws.handlers.RequestHandler2;
+import io.opentelemetry.javaagent.instrumentation.awssdk.v1_11.ContextScopePair;
+import io.opentelemetry.javaagent.tooling.TypeInstrumentation;
+import java.util.Map;
+import net.bytebuddy.asm.Advice;
+import net.bytebuddy.description.method.MethodDescription;
+import net.bytebuddy.description.type.TypeDescription;
+import net.bytebuddy.matcher.ElementMatcher;
+
+/**
+ * This is additional 'helper' to catch cases when HTTP request throws exception different from
+ * {@link AmazonClientException} (for example an error thrown by another handler). In these cases
+ * {@link RequestHandler2#afterError} is not called.
+ */
+public class AwsHttpClientInstrumentation implements TypeInstrumentation {
+
+  @Override
+  public ElementMatcher<TypeDescription> typeMatcher() {
+    return named("com.amazonaws.http.AmazonHttpClient");
+  }
+
+  @Override
+  public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
+    return singletonMap(
+        isMethod().and(not(isAbstract())).and(named("doExecute")),
+        AwsHttpClientInstrumentation.class.getName() + "$HttpClientAdvice");
+  }
+
+  public static class HttpClientAdvice {
+    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
+    public static void methodExit(
+        @Advice.Argument(value = 0, optional = true) Request<?> request,
+        @Advice.Thrown Throwable throwable) {
+      if (throwable != null) {
+        ContextScopePair scope = request.getHandlerContext(CONTEXT_SCOPE_PAIR_CONTEXT_KEY);
+        if (scope != null) {
+          request.addHandlerContext(CONTEXT_SCOPE_PAIR_CONTEXT_KEY, null);
+          tracer().endExceptionally(scope.getContext(), throwable);
+          scope.closeScope();
+        }
+      }
+    }
+  }
+}
