+++ instrumentation/aws-sdk/aws-sdk-1.11-core/javaagent/src/test/groovy/Aws1ClientTest.groovy
