+++ instrumentation/aws-sdk/aws-sdk-1.11-core/javaagent/src/test_before_1_11_106/groovy/Aws0ClientTest.groovy
