+++ instrumentation/aws-sdk/aws-sdk-1.11/library/aws-sdk-1.11-library.gradle
@@ -1,0 +1,5 @@
+apply from: "$rootDir/gradle/instrumentation-library.gradle"
+
+dependencies {
+  library group: 'com.amazonaws', name: 'aws-java-sdk-core', version: '1.11.0'
+}
