+++ instrumentation/aws-sdk/aws-sdk-1.11/library/src/main/java/io/opentelemetry/javaagent/instrumentation/awssdk/v1_11/ContextScopePair.java
