+++ settings.gradle
@@ -67,7 +67,9 @@ include ':instrumentation:async-http-client-1.9:javaagent'
 include ':instrumentation:aws-lambda-1.0:javaagent'
 include ':instrumentation:aws-lambda-1.0:library'
 include ':instrumentation:aws-lambda-1.0:testing'
-include ':instrumentation:aws-sdk:aws-sdk-1.11:javaagent'
+include ':instrumentation:aws-sdk:aws-sdk-1.11:library'
+include ':instrumentation:aws-sdk:aws-sdk-1.11-core:javaagent'
+include ':instrumentation:aws-sdk:aws-sdk-1.11-sqs:javaagent'
 include ':instrumentation:aws-sdk:aws-sdk-2.2:javaagent'
 include ':instrumentation:aws-sdk:aws-sdk-2.2:library'
 include ':instrumentation:aws-sdk:aws-sdk-2.2:testing'
